+++ docs/CHANGELOG.rst
@@ -23,6 +23,7 @@ Changed
   descriptor
 - Add ``annotations`` property to the ``Entity`` object in Python processes
 - Allow filtering ``AnnotationField`` by ids
+- Allow filtering ``AnnotationValue`` by full path
 
 
 ===================